AI-Projects/Day_06_Document_Validation_Agent/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=1.10
